// apps/frontend/src/api/generated/model/balancePoint.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.49
 */

export interface BalancePoint {
  event_date: string
  account_id: string
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  balance: string
  currency: string
}